+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # flitz
 
+## 0.5.0
+
+* `options` can also be an array of functions now
+* type checks of `options` parameter
+
 ## 0.4.1
 
 * [BREAKING CHANGE]: `RequestPathValidator` type gets an `IncomingMessage` object as argument instead a string now